Add toggleRead prototype function
Clicking on a checkmark of a newly added book to myLibrary will trigger a toggle of the object
in the myLibrary array. However it does not update on DOM to reflect the change visually
to the user. Also there's not persistent button for toggling read or unread on the cards yet.